+++ code/Samples/Gfx/Instancing/Instancing.cc
@@ -173,8 +173,8 @@ InstancingApp::OnInit() {
     this->drawState = Gfx::CreateResource(dss);
     
     // setup projection and view matrices
-    const float32 fbWidth = Gfx::DisplayAttrs().FramebufferWidth;
-    const float32 fbHeight = Gfx::DisplayAttrs().FramebufferHeight;
+    const float32 fbWidth = (const float32) Gfx::DisplayAttrs().FramebufferWidth;
+    const float32 fbHeight = (const float32) Gfx::DisplayAttrs().FramebufferHeight;
     this->proj = glm::perspectiveFov(glm::radians(45.0f), fbWidth, fbHeight, 0.01f, 100.0f);
     this->modelViewProj = this->proj * this->view * this->model;
