+++ code/Samples/Gfx/SimpleRenderTarget/SimpleRenderTarget.cc
@@ -96,7 +96,7 @@ SimpleRenderTargetApp::OnInit() {
         .Add(VertexAttr::Position, VertexFormat::Float3)
         .Add(VertexAttr::Normal, VertexFormat::Byte4N)
         .Add(VertexAttr::TexCoord0, VertexFormat::Float2);
-    shapeBuilder.Sphere(0.5f, 72.0f, 40.0f).Build();
+    shapeBuilder.Sphere(0.5f, 72, 40).Build();
     GfxId sphere = Gfx::CreateResource(MeshSetup::FromStream(), shapeBuilder.Result());
 
     // create shaders
@@ -114,8 +114,8 @@ SimpleRenderTargetApp::OnInit() {
     this->displayDrawState = Gfx::CreateResource(dispdsSetup);
     
     // setup static transform matrices
-    float32 fbWidth = Gfx::DisplayAttrs().FramebufferWidth;
-    float32 fbHeight = Gfx::DisplayAttrs().FramebufferHeight;
+    float32 fbWidth = (const float32) Gfx::DisplayAttrs().FramebufferWidth;
+    float32 fbHeight = (const float32) Gfx::DisplayAttrs().FramebufferHeight;
     this->offscreenProj = glm::perspective(glm::radians(45.0f), 1.0f, 0.01f, 20.0f);
     this->displayProj = glm::perspectiveFov(glm::radians(45.0f), fbWidth, fbHeight, 0.01f, 100.0f);
     this->view = glm::mat4();
